refactor: period 통일 및 패키지 이동
