+++ android/app/src/main/java/com/created/team201/data/mapper/StudyManagerMapper.kt
@@ -1,10 +1,10 @@
 package com.created.team201.data.mapper
 
-import com.created.domain.Period
-import com.created.domain.PeriodUnit
-import com.created.domain.StudySummary
+import com.created.domain.model.Period
+import com.created.domain.model.PeriodUnit
 import com.created.domain.model.Role
 import com.created.domain.model.StudyManage
+import com.created.domain.model.StudySummary
 import com.created.team201.data.remote.response.StudyManageResponseDto
 
 fun StudyManageResponseDto.toDomain(): StudyManage =
