+++ android/app/src/main/java/com/created/team201/data/repository/StudyListRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.created.team201.data.repository
 
-import com.created.domain.StudySummary
+import com.created.domain.model.StudySummary
 import com.created.domain.repository.StudyListRepository
 import com.created.team201.data.datasource.remote.StudyListDataSource
 import com.created.team201.data.mapper.toDomain
