+++ android/app/src/main/java/com/created/team201/presentation/createStudy/model/PeriodUiModel.kt
@@ -1,6 +1,8 @@
 package com.created.team201.presentation.createStudy.model
 
+import com.created.domain.model.PeriodUnit
+
 data class PeriodUiModel(
     val date: Int,
-    val type: Int,
+    val unit: PeriodUnit,
 )
