+++ android/app/src/main/java/com/created/team201/presentation/studyList/model/PeriodUiModel.kt
@@ -1,6 +1,6 @@
 package com.created.team201.presentation.studyList.model
 
-import com.created.domain.PeriodUnit
+import com.created.domain.model.PeriodUnit
 
 data class PeriodUiModel(
     val number: Int,
