--- android/domain/src/main/java/com/created/domain/Period.kt
@@ -1,6 +1,0 @@
-package com.created.domain
-
-data class Period(
-    val number: Int,
-    val unit: PeriodUnit
-)
