+++ android/domain/src/main/java/com/created/domain/model/Page.kt
@@ -1,4 +1,4 @@
-package com.created.domain
+package com.created.domain.model
 
 data class Page(val index: Int = 0) {
     operator fun inc(): Page {
