+++ android/domain/src/main/java/com/created/domain/model/Period.kt
@@ -2,5 +2,5 @@ package com.created.domain.model
 
 data class Period(
     val date: Int,
-    val type: Int,
+    val unit: PeriodUnit,
 )
