+++ android/domain/src/main/java/com/created/domain/model/StudyManage.kt
@@ -1,7 +1,5 @@
 package com.created.domain.model
 
-import com.created.domain.StudySummary
-
 data class StudyManage(
     val studySummary: StudySummary,
     val role: Role,
