+++ android/domain/src/main/java/com/created/domain/model/StudySummary.kt
@@ -1,4 +1,4 @@
-package com.created.domain
+package com.created.domain.model
 
 data class StudySummary(
     val id: Long,
